+++ src/option.h
@@ -1,5 +1,5 @@
 //
-//  Copyright (C) 2022-2023  Nick Gasson
+//  Copyright (C) 2022-2024  Nick Gasson
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -68,6 +68,7 @@ typedef enum {
    OPT_VHPI_DEBUG,
    OPT_SERVER_PORT,
    OPT_STDERR_LEVEL,
+   OPT_CHECK_SYNTHESIS,
 
    OPT_LAST_NAME
 } opt_name_t;